+++ chainstate/src/detail/chainstateref/mod.rs
@@ -35,7 +35,6 @@ use common::{
     time_getter::TimeGetterFn,
     Uint256,
 };
-use consensus::pos::is_due_for_epoch_data_calculation;
 use consensus::TransactionIndexHandle;
 use logging::log;
 use pos_accounting::{FlushablePoSAccountingView, PoSAccountingDB, PoSAccountingDeltaData};
@@ -768,7 +767,7 @@ impl<'a, S: BlockchainStorageWrite, O: OrphanBlocksMut, V: TransactionVerificati
         // Connect the new chain
         for block_index in new_chain {
             self.connect_tip(&block_index).log_err()?;
-            self.activate_seal_epoch(block_index.block_height())?;
+            self.activate_epoch_seal(block_index.block_height())?;
         }
 
         Ok(())
@@ -957,8 +956,8 @@ impl<'a, S: BlockchainStorageWrite, O: OrphanBlocksMut, V: TransactionVerificati
         }
     }
 
-    fn activate_seal_epoch(&mut self, height: BlockHeight) -> Result<(), BlockError> {
-        if is_due_for_epoch_data_calculation(self.chain_config, height) {
+    fn activate_epoch_seal(&mut self, height: BlockHeight) -> Result<(), BlockError> {
+        if self.chain_config.is_due_for_epoch_seal(height) {
             let current_epoch_index = self.chain_config.epoch_index_from_height(&height);
             let sealed_epoch_distance_from_tip =
                 self.chain_config.sealed_epoch_distance_from_tip() as u64;
@@ -966,10 +965,9 @@ impl<'a, S: BlockchainStorageWrite, O: OrphanBlocksMut, V: TransactionVerificati
             if current_epoch_index >= sealed_epoch_distance_from_tip {
                 let epoch_index_to_seal = current_epoch_index - sealed_epoch_distance_from_tip;
 
-                let first_block_epoch_to_seal =
-                    epoch_index_to_seal * self.chain_config.epoch_length().get();
-                let last_block_epoch_to_seal =
-                    first_block_epoch_to_seal + self.chain_config.epoch_length().get() - 1;
+                let epoch_length = self.chain_config.epoch_length().get();
+                let first_block_epoch_to_seal = epoch_index_to_seal * epoch_length;
+                let last_block_epoch_to_seal = first_block_epoch_to_seal + epoch_length - 1;
 
                 // iterate over every block in the epoch and merge every block delta into a singe delta
                 let epoch_delta = (first_block_epoch_to_seal..=last_block_epoch_to_seal)
