+++ chainstate/storage/src/internal/mod.rs
@@ -372,7 +372,7 @@ impl<B: storage::Backend> BlockchainStorageWrite for Store<B> {
         ) -> crate::Result<()>;
         fn del_accounting_undo_data(&mut self, id: Id<Block>) -> crate::Result<()>;
 
-        fn set_accounting_delta(
+        fn apply_accounting_delta(
             &mut self,
             id: Id<Block>,
             delta: &PoSAccountingDeltaData,
