+++ chainstate/storage/src/lib.rs
@@ -179,7 +179,7 @@ pub trait BlockchainStorageWrite:
     fn del_accounting_undo_data(&mut self, id: Id<Block>) -> crate::Result<()>;
 
     // Set accounting delta for specific block
-    fn set_accounting_delta(
+    fn apply_accounting_delta(
         &mut self,
         id: Id<Block>,
         delta: &PoSAccountingDeltaData,
