+++ chainstate/tx-verifier/src/transaction_verifier/tests/mock.rs
@@ -112,7 +112,7 @@ mockall::mock! {
             tx_source: TransactionSource,
         ) -> Result<(), TransactionVerifierStorageError>;
 
-        fn set_accounting_delta(
+        fn apply_accounting_delta(
             &mut self,
             tx_source: TransactionSource,
             delta: &PoSAccountingDeltaData,
