--- consensus/src/pos/mod.rs
@@ -1,21 +1,0 @@
-// Copyright (c) 2023 RBB S.r.l
-// opensource@mintlayer.org
-// SPDX-License-Identifier: MIT
-// Licensed under the MIT License;
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// https://github.com/mintlayer/mintlayer-core/blob/master/LICENSE
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use common::{chain::ChainConfig, primitives::BlockHeight};
-
-pub fn is_due_for_epoch_data_calculation(chain_config: &ChainConfig, height: BlockHeight) -> bool {
-    let height: u64 = height.into();
-    height % chain_config.epoch_length() == 0
-}
